cifar/requirements.txt
absl-py==2.1.0
array_record==0.5.1
asttokens==2.4.1
astunparse==1.6.3
certifi==2024.8.30
charset-normalizer==3.3.2
chex==0.1.86
click==8.1.7
comm==0.2.2
contextlib2==21.6.0
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
diffrax==0.6.0
dm-tree==0.1.8
docker-pycreds==0.4.0
docstring_parser==0.16
equinox==0.11.6
etils==1.9.4
exceptiongroup==1.2.2
executing==2.1.0
flatbuffers==24.3.25
flax==0.9.0
fonttools==4.53.1
fsspec==2024.9.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.66.1
h5py==3.11.0
humanize==4.10.0
idna==3.10
immutabledict==4.2.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.27.0
jax==0.4.28
jax-cuda12-pjrt==0.4.28
jax-cuda12-plugin==0.4.28
jaxlib==0.4.28+cuda12.cudnn89
jaxtyping==0.2.34
jedi==0.19.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.5.0
kiwisolver==1.4.7
libclang==18.1.1
lineax==0.0.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
ml_collections==0.1.1
msgpack==1.1.0
namex==0.0.8
nest-asyncio==1.6.0
numpy==1.26.4
nvidia-cublas-cu12==12.6.1.4
nvidia-cuda-cupti-cu12==12.6.68
nvidia-cuda-nvcc-cu12==12.6.68
nvidia-cuda-nvrtc-cu12==12.6.68
nvidia-cuda-runtime-cu12==12.6.68
nvidia-cudnn-cu12==8.9.7.29
nvidia-cufft-cu12==11.2.6.59
nvidia-cusolver-cu12==11.6.4.69
nvidia-cusparse-cu12==12.5.3.3
nvidia-nccl-cu12==2.23.4
nvidia-nvjitlink-cu12==12.6.68
opt-einsum==3.3.0
optax==0.2.3
optimistix==0.0.7
optree==0.12.1
orbax-checkpoint==0.6.4
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.4
promise==2.3
prompt_toolkit==3.0.47
protobuf==3.20.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
rich==13.8.1
scipy==1.14.1
sentry-sdk==2.14.0
setproctitle==1.3.3
simple-parsing==0.1.6
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-datasets==4.9.6
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.15.0
tensorstore==0.1.65
termcolor==2.4.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typeguard==2.13.3
typing_extensions==4.12.2
urllib3==2.2.3
wandb==0.18.1
wcwidth==0.2.13
Werkzeug==3.0.4
wrapt==1.16.0
zipp==3.20.2
